requirements.txt
appdirs==1.4.0
Django==1.10
django-cors-headers==2.0.2
django-filter==1.0.1
djangorestframework==3.5.4
Jinja2==2.8
Markdown==2.6.8
MarkupSafe==0.23
packaging==16.8
pyparsing==2.1.10
six==1.10.0
